Migrate goStalkService to Typescript
